switch from using int to long in IDataProvider and the cascade from that Issue: WICKET-1175
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1203891 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket-core/src/main/java/org/apache/wicket/markup/html/list/Loop.java
@@ -49,9 +49,9 @@ public abstract class Loop extends AbstractRepeater
 	 *            max index of the loop
 	 * @see org.apache.wicket.Component#Component(String, IModel)
 	 */
-	public Loop(final String id, final int iterations)
+	public Loop(final String id, final long iterations)
 	{
-		super(id, new Model<Integer>(iterations));
+		super(id, new Model<Long>(iterations));
 	}
 
 	/**
@@ -63,7 +63,7 @@ public abstract class Loop extends AbstractRepeater
 	 *            Must contain a Integer model object
 	 * @see org.apache.wicket.Component#Component(String, IModel)
 	 */
-	public Loop(final String id, final IModel<Integer> model)
+	public Loop(final String id, final IModel<Long> model)
 	{
 		super(id, model);
 	}
